requirements.txt
Django==2.1.11
django-bootstrap-modal-forms==1.4.1
django-widget-tweaks==1.4.2
selenium==3.14.0
pytz==2018.5

atomicwrites       ==          1.3.0
attrs            ==            19.3.0
Babel          ==              2.8.0
bson         ==                0.5.8
certifi    ==                ==  2020.4.5.1
chardet                  ==    3.0.4
colorama               ==      0.4.3
coverage             ==        5.1
dataclasses               ==   0.6
Django                  ==     2.1.11
django-bootstrap-modal-forms== 1.4.1
django-classy-tags    ==       1.0.0
django-cms                ==   3.7.1
django-extensions       ==     2.2.9
django-filer          ==       1.7.0
django-formtools    ==         2.2
django-jquery-js      ==       3.1.1
django-js-asset     ==         1.2.2
django-mptt       ==           0.11.0
django-phonenumber-field ==    4.0.0
django-polymorphic ==          2.1.2
django-sekizai   ==            1.1.0
django-treebeard           ==  4.3.1
django-widget-tweaks     ==    1.4.2
djangocms-admin-style  ==      1.5.0
djangorestframework  ==        3.11.0
djongo             ==          1.3.1
easy-thumbnails         ==     2.7
idna                  ==       2.9
importlib-metadata  ==         1.6.0
more-itertools    ==           8.2.0
packaging       ==             20.3
phonenumbers  ==               8.12.1
Pillow      ==                 7.1.1
pip       ==                   20.0.2
pluggy                  ==     0.13.1
py                    ==       1.8.1
PyJWT               ==         1.7.1
pymongo           ==           3.10.1
pyparsing       ==             2.4.7
pytest        ==               5.4.1
pytest-cov  ==                 2.8.1
pytest-django               == 3.9.0
python-dateutil           ==   2.8.1
pytz                    ==     2018.5
requests              ==       2.23.0
selenium            ==         3.14.0
setuptools         ==          46.1.3
six              ==            1.14.0
sqlparse       ==              0.2.4
twilio           ==            6.38.0
Unidecode      ==              1.1.1
urllib3                      ==1.25.8
wcwidth                    ==  0.1.9
wheel                    ==    0.34.2
whitenoise             ==      5.0.1
zipp ==   3.1.0

